+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Translation records with a negative weight will not be used when a fallback is given
 * When the parameters contain a `fallback` property, it will be used as the fallback value and not as a filter
 * If there are no translation records with a positive filter score, they will not be used
+* Ignore parameters that have a `null` or `undefined` value
 
 ## 0.6.8 (2024-01-22)